kNN classifier for Income datset
This piece of code uses the kNN classifier to classify the test records
into one of two classes. The code consists of logic for both the Iris
and Income dataset..

Python code is much faster now. Takes a maximum of 15 seconds to run on
the income_TRAIN_FINAL and income_TEST_FINAL csv files